hashcat-cli.c: fix build against musl libc
GLIBC's termio.h is just an empty file with 2 include directives:

 #include <termios.h>
 #include <sys/ioctl.h>

and a comment that says that it is obsolete and you should use termios.h
and struct termios instead - exactly what musl libc does.

since the obsolete interface was replaced with termios by POSIX, musl
only implements the standardized interface.

also, instead of doing raw ioctl syscalls, we now use POSIX' tcgetattr()
and tcsetattr(), which are the high-level API to deal with struct termios.